Secured endpoints

+++ src/main/java/hu/elte/absencetracker/controllers/AbsenceController.java
@@ -1,7 +1,9 @@
 package hu.elte.absencetracker.controllers;
 
 import hu.elte.absencetracker.entities.Absence;
+import hu.elte.absencetracker.entities.User;
 import hu.elte.absencetracker.repositories.AbsenceRepository;
+import hu.elte.absencetracker.security.AuthenticatedUser;
 import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -17,47 +19,80 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequestMapping("/absences")
 public class AbsenceController {
-    
+
     @Autowired
     private AbsenceRepository absenceRepository;
-    
+
+    @Autowired
+    private AuthenticatedUser authenticatedUser;
+
     @GetMapping("")
     public ResponseEntity<Iterable<Absence>> getAll() {
-        return ResponseEntity.ok(absenceRepository.findAll());
+        User authUser = authenticatedUser.getUser();
+        if (authUser.getRole().equals("ADMIN")) {
+            return ResponseEntity.ok(absenceRepository.findAll());
+        } else {
+            return ResponseEntity.badRequest().build();
+        }
+
     }
-    
+
     @GetMapping("/{id}")
     public ResponseEntity<Absence> get(@PathVariable Integer id) {
         Optional<Absence> absence = absenceRepository.findById(id);
-        if(absence.isPresent()) {
-            return ResponseEntity.ok(absence.get());
+        User authUser = authenticatedUser.getUser();
+        if (absence.isPresent()) {
+            if (authUser.getRole().equals("ADMIN")) {
+                return ResponseEntity.ok(absence.get());
+            } else {
+                return ResponseEntity.badRequest().build();
+            }
+
         } else {
             return ResponseEntity.notFound().build();
         }
     }
-    
+
     @PostMapping("")
     public ResponseEntity<Absence> post(@RequestBody Absence absence) {
-        return ResponseEntity.ok(absenceRepository.save(absence));
+        User authUser = authenticatedUser.getUser();
+        if (authUser.getRole().equals("ADMIN")) {
+            return ResponseEntity.ok(absenceRepository.save(absence));
+        } else {
+            return ResponseEntity.badRequest().build();
+        }
+
     }
-    
+
     @PutMapping("/{id}")
     public ResponseEntity<Absence> update(@PathVariable Integer id, @RequestBody Absence newAbsence) {
         Optional<Absence> oldAbsence = absenceRepository.findById(id);
-        if(oldAbsence.isPresent()) {
-            newAbsence.setId(id);
-            return ResponseEntity.ok(absenceRepository.save(newAbsence));
+        User authUser = authenticatedUser.getUser();
+        if (oldAbsence.isPresent()) {
+            if (authUser.getRole().equals("ADMIN")) {
+                newAbsence.setId(id);
+                return ResponseEntity.ok(absenceRepository.save(newAbsence));
+            } else {
+                return ResponseEntity.badRequest().build();
+            }
+
         } else {
             return ResponseEntity.notFound().build();
         }
     }
-    
+
     @DeleteMapping("/{id}")
     public ResponseEntity<Absence> delete(@PathVariable Integer id) {
         Optional<Absence> absence = absenceRepository.findById(id);
-        if(absence.isPresent()) {
-            absenceRepository.deleteById(id);
-            return ResponseEntity.ok().build();
+        User authUser = authenticatedUser.getUser();
+        if (absence.isPresent()) {
+            if (authUser.getRole().equals("ADMIN")) {
+                absenceRepository.deleteById(id);
+                return ResponseEntity.ok().build();
+            } else {
+                return ResponseEntity.badRequest().build();
+            }
+
         } else {
             return ResponseEntity.notFound().build();
         }
