+++ src/main/java/hu/elte/absencetracker/controllers/SubjectController.java
@@ -2,8 +2,10 @@ package hu.elte.absencetracker.controllers;
 
 import hu.elte.absencetracker.entities.Lesson;
 import hu.elte.absencetracker.entities.Subject;
+import hu.elte.absencetracker.entities.User;
 import hu.elte.absencetracker.repositories.LessonRepository;
 import hu.elte.absencetracker.repositories.SubjectRepository;
+import hu.elte.absencetracker.security.AuthenticatedUser;
 import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -26,6 +28,9 @@ public class SubjectController {
     @Autowired
     private LessonRepository lessonRepository;
     
+    @Autowired
+    private AuthenticatedUser authenticatedUser;
+    
     @GetMapping("")
     public ResponseEntity<Iterable<Subject>> getAll() {
         return ResponseEntity.ok(subjectRepository.findAll());
@@ -43,15 +48,26 @@ public class SubjectController {
     
     @PostMapping("")
     public ResponseEntity<Subject> post(@RequestBody Subject subject) {
-        return ResponseEntity.ok(subjectRepository.save(subject));
+        User authUser = authenticatedUser.getUser();
+        if(authUser.getRole().equals("ADMIN")){
+            return ResponseEntity.ok(subjectRepository.save(subject));
+        } else{
+            return ResponseEntity.badRequest().build();
+        }
+        
     }
     
     @PutMapping("/{id}")
     public ResponseEntity<Subject> update(@PathVariable Integer id, @RequestBody Subject newSubject) {
         Optional<Subject> oldSubject = subjectRepository.findById(id);
+        User authUser = authenticatedUser.getUser();
         if(oldSubject.isPresent()) {
-            newSubject.setId(id);
-            return ResponseEntity.ok(subjectRepository.save(newSubject));
+            if(authUser.getRole().equals("ADMIN")){
+                newSubject.setId(id);
+                return ResponseEntity.ok(subjectRepository.save(newSubject));
+            } else{
+                return ResponseEntity.badRequest().build();
+            }
         } else {
             return ResponseEntity.notFound().build();
         }
@@ -60,9 +76,15 @@ public class SubjectController {
     @DeleteMapping("/{id}")
     public ResponseEntity<Subject> delete(@PathVariable Integer id) {
         Optional<Subject> subject = subjectRepository.findById(id);
+        User authUser = authenticatedUser.getUser();
         if(subject.isPresent()) {
-            subjectRepository.deleteById(id);
-            return ResponseEntity.ok().build();
+            if(authUser.getRole().equals("ADMIN")){
+                subjectRepository.deleteById(id);
+                return ResponseEntity.ok().build();
+            } else{
+                 return ResponseEntity.badRequest().build();
+            }
+            
         } else {
             return ResponseEntity.notFound().build();
         }
@@ -81,9 +103,15 @@ public class SubjectController {
     @PostMapping("/{id}/lessons")
     public ResponseEntity<Lesson> postLesson(@PathVariable Integer id, @RequestBody Lesson lesson) {
         Optional<Subject> subject = subjectRepository.findById(id);
+        User authUser = authenticatedUser.getUser();
         if(subject.isPresent()) {
-            lesson.setSubject(subject.get());
-            return ResponseEntity.ok(lessonRepository.save(lesson));
+            if(authUser.getRole().equals("ADMIN")){
+                 lesson.setSubject(subject.get());
+                return ResponseEntity.ok(lessonRepository.save(lesson));
+            } else{
+                return ResponseEntity.badRequest().build();
+            }
+            
         } else {
             return ResponseEntity.notFound().build();
         }
